pyproject.toml
[tool.poetry]
name = "agent-development"
version = "0.1.0"
description = "AI workflow and agent patterns"
authors = ["Shervin <shervin.khazaeli@yahoo.com>"]
packages = [
    { include = "agent_dev", from = "src" },
    { include = "water_industry", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.11.1"
langgraph = "^0.4.1"
ipykernel = "^6.29.0"
jupyter = "^1.1.1"
langchain-core = "^0.3.58"
langchain-anthropic = "^0.3.12"
python-dotenv = "^1.1.0"
langchain = "^0.3.25"
openai = "^1.77.0"
matplotlib = "^3.8.2"
pandas = "^2.1.3"
reportlab = "^4.0.7" 


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.1.0"
black = "^24.3.0"
ruff = "^0.4.3"
pytest = "^8.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
